Revert "WIP refactoring into separate modules - should prob do this in another PR"
This reverts commit c8d4470840f80a849c231cf873231ffa1e4afc21.
