--- mlflow/projects/execution.py
@@ -1,3 +1,0 @@
-"""
-Module containing methods for running projects. Imported after other stuff.
-"""
