--- mlflow/projects/utils.py
@@ -1,96 +1,0 @@
-from __future__ import print_function
-
-import hashlib
-import re
-import os
-import shutil
-import tempfile
-
-from distutils import dir_util
-
-from mlflow.projects._project_spec import Project
-from mlflow.utils import file_utils, process
-
-
-_GIT_URI_REGEX = re.compile(r"^[^/]*:")
-
-class ExecutionException(Exception):
-    """Exception thrown when executing a project fails."""
-    pass
-
-
-def _load_project(work_dir, uri):
-    return Project(_expand_uri(uri), file_utils.read_yaml(work_dir, "MLproject"))
-
-
-def _get_conda_env_name(conda_env_path):
-    with open(conda_env_path) as conda_env_file:
-        conda_env_hash = hashlib.sha1(conda_env_file.read().encode("utf-8")).hexdigest()
-    return "mlflow-%s" % conda_env_hash
-
-
-def _fetch_git_repo(uri, version, dst_dir, git_username, git_password):
-    """
-    Clones the git repo at `uri` into `dst_dir`, checking out commit `version` (or defaulting
-    to the head commit of the repository's master branch if version is unspecified). If git_username
-    and git_password are specified, uses them to authenticate while fetching the repo. Otherwise,
-    assumes authentication parameters are specified by the environment, e.g. by a Git credential
-    helper.
-    """
-    # We defer importing git until the last moment, because the import requires that the git
-    # executable is availble on the PATH, so we only want to fail if we actually need it.
-    import git
-    repo = git.Repo.init(dst_dir)
-    origin = repo.create_remote("origin", uri)
-    git_args = [git_username, git_password]
-    if not (all(arg is not None for arg in git_args) or all(arg is None for arg in git_args)):
-        raise ExecutionException("Either both or neither of git_username and git_password must be "
-                                 "specified.")
-    if git_username:
-        git_credentials = "url=%s\nusername=%s\npassword=%s" % (uri, git_username, git_password)
-        repo.git.config("--local", "credential.helper", "cache")
-        process.exec_cmd(cmd=["git", "credential-cache", "store"], cwd=dst_dir,
-                         cmd_stdin=git_credentials)
-    origin.fetch()
-    if version is not None:
-        repo.git.checkout(version)
-    else:
-        repo.create_head("master", origin.refs.master)
-        repo.heads.master.checkout()
-
-
-def _fetch_project(uri, version, dst_dir, git_username, git_password):
-    """Download a project to the target `dst_dir` from a Git URI or local path."""
-    if _GIT_URI_REGEX.match(uri):
-        # Use Git to clone the project
-        _fetch_git_repo(uri, version, dst_dir, git_username, git_password)
-    else:
-        if version is not None:
-            raise ExecutionException("Setting a version is only supported for Git project URIs")
-        # TODO: don't copy mlruns directory here
-        # Note: uri might be equal to dst_dir, e.g. if we're not using a temporary work dir
-        if uri != dst_dir:
-            dir_util.copy_tree(src=uri, dst=dst_dir)
-
-    # Make sure they don't have an outputs or mlruns directory (will need to change if we change
-    # how we log results locally)
-    shutil.rmtree(os.path.join(dst_dir, "outputs"), ignore_errors=True)
-    shutil.rmtree(os.path.join(dst_dir, "mlruns"), ignore_errors=True)
-
-
-def _get_work_dir(uri, use_temp_cwd):
-    """
-    Returns a working directory to use for fetching & running the project with the specified URI.
-    :param use_temp_cwd: Only used if `uri` is a local directory. If True, returns a temporary
-                         working directory.
-    """
-    if _GIT_URI_REGEX.match(uri) or use_temp_cwd:
-        # Create a temp directory to download and run the project in
-        return tempfile.mkdtemp(prefix="mlflow-")
-    return os.path.abspath(uri)
-
-
-def _expand_uri(uri):
-    if _GIT_URI_REGEX.match(uri):
-        return uri
-    return os.path.abspath(uri)
